Suggested refactor of the StateChangeAction impl;  Pull Request and JIRA incoming
