+++ notification-portlet-webapp/src/main/java/org/jasig/portlet/notice/service/CacheNotificationService.java
@@ -191,15 +191,4 @@ public final class CacheNotificationService extends AbstractNotificationService 
 
     }
 
-	/**
-	 * Clear cache for user associated with req.
-	 * 
-	 * @param req 
-	 */
-	public void clearCacheForUser(final PortletRequest req) {
-		if (usernameFinder.isAuthenticated(req)) {
-			final String cacheKey = createServiceUserWindowSpecificCacheKey(req);
-			cache.remove(cacheKey);
-		}
-	}
 }
